% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{any_button_creator}
\alias{any_button_creator}
\title{Function to create special, user-defined buttons to be rendered in a Data Table.}
\usage{
any_button_creator(.button_id, .button_label, index)
}
\arguments{
\item{.button_id}{The ID to use for the button. Will not be namespaced to the module.}

\item{.button_label}{The label to display on the button.}

\item{index}{The index of the Data Table row that the button will be placed in.}
}
\description{
Function to create special, user-defined buttons to be rendered in a Data Table.
}
