% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{api_get}
\alias{api_get}
\title{submit a Get request to the RSQ API}
\usage{
api_get(endpoint_url, ..., return_df = FALSE)
}
\arguments{
\item{endpoint_url}{the full API URL for the endpoint to receive the GET request}

\item{...}{query parameters to be sent with the API request.}

\item{return_df}{boolean to indicate if the response content should be converted to a DF via the "simplify_df" option of jsonlite::fromJSON.}
}
\value{
a list-object (if return_df is false) or data.frame (if return_df is true) containing the HTTP response content.
}
\description{
submit a Get request to the RSQ API
}
