% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{api_post}
\alias{api_post}
\title{submit a POST request to the RSQ API}
\usage{
api_post(endpoint_url, body = NULL, ..., return_df = FALSE)
}
\arguments{
\item{endpoint_url}{the full API URL for the endpoint to receive the GET request}

\item{body}{a list object that will be converted into a JSON body and attached to the REST query.}

\item{...}{query parameters sent with the API request.}

\item{return_df}{boolean to indicate if the response content should be converted to a DF via the "simplify_df" option of jsonlite::fromJSON.}
}
\value{
a list-object (if return_df is false) or data.frame (if return_df is true) containing the HTTP response content.
}
\description{
submit a POST request to the RSQ API
}
