% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{cancel_button_creator}
\alias{cancel_button_creator}
\title{Function to create a cancel button to be rendered in a Data Table.}
\usage{
cancel_button_creator(index, .ns)
}
\arguments{
\item{index}{The index of the Data Table row that the button will be placed in.}

\item{.ns}{The namespace function that should be applied to the button's ID.}
}
\description{
Function to create a cancel button to be rendered in a Data Table.
}
\details{
The function returns the cancel button as raw HTML with a JavaScript statement
embedded. Using raw HTML + JavaScript allows us to embed the button directly into
a DataTable object.
}
