% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{cancel_job}
\alias{cancel_job}
\title{cancel_job}
\usage{
cancel_job(job, api_config = getOption("rrsq.rsq_api_config"))
}
\arguments{
\item{job}{the numeric ID of the RSQ job to return, OR an object of class "rsq_job"}

\item{api_config}{an rsq_api_config object}
}
\value{
the response content from the cancel job API call.
}
\description{
function to hit the "cancel job" endpoint to cancel an RSQ job.
}
