% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{check_api_config}
\alias{check_api_config}
\title{check_api_config}
\usage{
check_api_config(api_config)
}
\arguments{
\item{api_config}{an api_config object}
}
\value{
returns true if the api_config object is valid, otherwise throws an error with rlang::abort
}
\description{
verifies that the given api_config object is a valid api_config object by ensuring that it has
the proper class attached to it.
}
