% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{check_response}
\alias{check_response}
\title{check_response}
\usage{
check_response(resp)
}
\arguments{
\item{resp}{an HTTP Response object}
}
\description{
function to check an HTTP response code and throw a relevant error, if applicable.
}
