% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{check_valid_statuses}
\alias{check_valid_statuses}
\title{check_valid_status}
\usage{
check_valid_statuses(statuses)
}
\arguments{
\item{statuses}{a case-sensitive character vector of statuses.}
}
\value{
TRUE if all statuses in statuses can be recognized by RSQ, otherwise throws an error with rlang::abort
}
\description{
verifies that all statuses in a vector of statuses can be recognized by RSQ.
}
