% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{collect_r_details}
\alias{collect_r_details}
\title{Information about running R}
\usage{
collect_r_details(.envs = NULL, .pkgs = NULL)
}
\arguments{
\item{.envs}{environment variables to retain}

\item{.pkgs}{packages to load in new session, if null defaults to all loaded in current session}
}
\value{
list with r_path, renv (list with set environment variables), work_dir
}
\description{
Information about running R
}
\details{
the resulting environment variables will always have R_LIBS_SITE,
R_LIBS_USER, and R_LIBS set to the .libPaths() to forceably make
sure that the remote session will leverage the same libpaths
}
