% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONSTANTS}
\alias{CONSTANTS}
\title{CONSTANTS}
\format{
An object of class \code{list} of length 4.
}
\usage{
CONSTANTS
}
\description{
Helper object to store constants such as endpoint addresses and such.
}
\keyword{datasets}
