% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{create_api_config}
\alias{create_api_config}
\title{create_api_config}
\usage{
create_api_config(
  host = get_api_host(),
  port = get_api_port(),
  root = get_api_root(),
  test_ping = TRUE,
  test_ping_timeout = 10
)
}
\arguments{
\item{host}{the hostname of server hosting the desired RSQ API. Should include the HTTP/HTTPS schema.}

\item{port}{the port on which the API his hosted}

\item{root}{the root of the API, without leading or trailing slashes.}

\item{test_ping}{boolean value. If TRUE, will ping the RSQ Service to verify that is is responding.
Will throw an error if the service is not responsive.}

\item{test_ping_timeout}{an integer value (in seconds) determining how long the test_ping should
wait for the RSQ service to reply before determining that the service is unresponsive.}
}
\value{
an object of class "rsq_api_config" (set by CONSTANTS$CLASSES$API_CONFIG) that can be passed
into other functions to facilitate the submission of API queries.
}
\description{
function to create a top level "API config" object, which stores basic, root-level information
about the specific RSQ API to be queried. The API config object can be used to construct endpoints,
as well as to store authorization information (if this becomes applicable in the future)
}
