% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-job.R
\name{.submit_job}
\alias{.submit_job}
\title{submit job}
\usage{
.submit_job(srvr, user, context, rscript_path, work_dir = NULL,
  r_path = NULL, .envs = NULL, .no_submit = FALSE)
}
\arguments{
\item{srvr}{rsq server url}

\item{user}{user submitting job}

\item{context}{job context}

\item{rscript_path}{path to rscript}

\item{work_dir}{directory to start R from}

\item{r_path}{path to R executable}

\item{.envs}{vector of env variables}

\item{.no_submit}{return json that would be submitted but don't make request}
}
\description{
submit job
}
