% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{extract_json}
\alias{extract_json}
\title{extract_json
function to pull JSON out of a response object.}
\usage{
extract_json(response, ..., simplify_df = FALSE)
}
\arguments{
\item{response}{the response object, which should have a $content field that is just raw JSON.}

\item{...}{additional arguments to fromJSON.}
}
\description{
extract_json
function to pull JSON out of a response object.
}
