% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{get_api_host}
\alias{get_api_host}
\title{get_api_host}
\usage{
get_api_host(host = Sys.getenv("RSQ_API_HOST"))
}
\arguments{
\item{host}{the host to use, if not the default.}
}
\description{
function to get the API host.
}
