% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{get_api_port}
\alias{get_api_port}
\title{get_api_port}
\usage{
get_api_port(port = as.numeric(Sys.getenv("RSQ_API_PORT")))
}
\arguments{
\item{port}{the port to use, if not the default}
}
\description{
function to get the API port
}
