% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{get_api_root}
\alias{get_api_root}
\title{get_api_root}
\usage{
get_api_root(root = Sys.getenv("RSQ_API_ROOT"))
}
\arguments{
\item{root}{the root to use, if not the default}
}
\description{
function to retrieve the API root (might not apply to RSQ)
}
