% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbing.R
\name{get_endpoint_url}
\alias{get_endpoint_url}
\title{get_endpoint_url}
\usage{
get_endpoint_url(api_config, endpoint)
}
\arguments{
\item{api_config}{an API Config object. Should be validated prior to passing into this function.}

\item{endpoint}{the endpoint to hit (string). Should not include leading/trailing /'s but should include inner /'s.}
}
\description{
function to get the usable URL to an endpoint
}
