% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{get_job}
\alias{get_job}
\title{get_job}
\usage{
get_job(job, return_df = FALSE, api_config = getOption("rrsq.rsq_api_config"))
}
\arguments{
\item{return_df}{boolean. If TRUE, results will be returned as a one-row data.frame.
Otherwise, results will be returned as a list-object.}

\item{api_config}{an api_config object}

\item{job_id}{the numeric ID of the RSQ job to return, OR an object of class "rsq_job"}
}
\description{
function to query an RSQ API and return the current status/info of single job.
job can be returned as a list object with class `rsq_job`, or as a data.frame.
}
