% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{get_jobs}
\alias{get_jobs}
\title{get_jobs}
\usage{
get_jobs(
  users = NULL,
  statuses = NULL,
  api_config = getOption("rrsq.rsq_api_config")
)
}
\arguments{
\item{users}{an optional parameter. If set, will only return jobs for the indicated users. Can be a string or character vector.}

\item{api_config}{an api_config object.}

\item{status}{an optional parameter. If set, will only return jobs containing the listed statuses.}
}
\value{
a data.frame of all current jobs.
}
\description{
function to query an RSQ API and return a data.frame of all jobs
}
