% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{get_valid_statuses}
\alias{get_valid_statuses}
\title{get_valid_statuses}
\usage{
get_valid_statuses()
}
\value{
a named list containing all valid RSQ statuses
}
\description{
helper function to return the current valid RSQ statuses
}
