% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{get_version}
\alias{get_version}
\title{get_version}
\usage{
get_version(api_config = getOption("rrsq.rsq_api_config"))
}
\arguments{
\item{api_config}{an rsq_api_config object}
}
\value{
a list object containing the version of the RSQ service.
}
\description{
function to return the version of the running RSQ service
}
