% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-jobs.R
\name{job_to_df}
\alias{job_to_df}
\title{convert a single job object into a dataframe.}
\usage{
job_to_df(.j)
}
\arguments{
\item{.j}{A single job object.}
}
\description{
convert a single job object into a dataframe.
}
