% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-jobs.R
\name{jobs_to_df}
\alias{jobs_to_df}
\title{convert list of jobs returned from db to a tidy dataframe}
\usage{
jobs_to_df(.jl)
}
\arguments{
\item{.jl}{job list}
}
\description{
convert list of jobs returned from db to a tidy dataframe
}
\details{
current drops some of the run details that shouldn't matter to the user
such as the script execution location
}
