% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-jobs.R
\name{map_dfr}
\alias{map_dfr}
\title{don't use purrr as requires dplyr for map_dfr}
\usage{
map_dfr(.x, .f, ...)
}
\arguments{
\item{.x}{A list to iterate over.}

\item{.f}{A function to apply to each element in the list.}

\item{...}{Additional arguments to be supplied to .f}
}
\description{
don't use purrr as requires dplyr for map_dfr
}
