% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{ping}
\alias{ping}
\title{ping}
\usage{
ping(timeout = 10, api_config = getOption("rrsq.rsq_api_config"))
}
\arguments{
\item{timeout}{parameter to determine how long (in seconds) the ping should wait for a response. Defaults to 10 seconds.}

\item{api_config}{an rsq_api_config object}
}
\value{
TRUE if the service is responsive, FALSE otherwise.
}
\description{
function to hit the "ping" endpoint of an RSQ service
in order to verify that the service is responding.
}
