% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{queueViewOutput}
\alias{queueViewOutput}
\title{queueViewOutput - Data Table displaying the jobs currently in an R Simple Queue.}
\usage{
queueViewOutput(id)
}
\arguments{
\item{id}{The Shiny ID to be associated with this object.}
}
\description{
queueViewOutput - Data Table displaying the jobs currently in an R Simple Queue.
}
