% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{set_api_config}
\alias{set_api_config}
\title{set_api_config}
\usage{
set_api_config(api_config)
}
\arguments{
\item{api_config}{an rsq_api_config object}
}
\value{
the newly set value of the rrsq.rsq_api_config option (i.e. api_config argument passed in)
}
\description{
Function to set the rrsq.rsq_api_config option to an api_config object.
doing this will allow you to call other functions without having to
pass in the api_config object everytime.
}
\details{
Note that this function _does_ verify that api_config is a valid rsq_api_config object.
As such, an error will be thrown if api_config is not a valid rsq_api_config object.
}
