% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{submit_job}
\alias{submit_job}
\title{submit job}
\usage{
submit_job(
  user,
  context,
  rscript_path,
  work_dir = NULL,
  r_path = NULL,
  renv = NULL,
  api_config = getOption("rrsq.rsq_api_config")
)
}
\arguments{
\item{user}{user submitting job}

\item{context}{job context}

\item{rscript_path}{path to rscript}

\item{work_dir}{directory to start R from}

\item{r_path}{path to R executable}

\item{renv}{list of environment variables}

\item{api_config}{an rsq_api_config object}

\item{srvr}{rsq server url}
}
\value{
an `rsq_job` object.
}
\description{
submit a job to RSQ using an rsq_api_config object.
}
