% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_difference}
\alias{time_difference}
\title{calculate the time difference}
\usage{
time_difference(end_time, start_time, units = "secs")
}
\arguments{
\item{end_time}{the end time}

\item{start_time}{the start time}

\item{units}{units to calculate, defaults to secs}
}
\description{
calculate the time difference
}
