% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-jobs.R
\name{to_datetime}
\alias{to_datetime}
\title{Function to format an object into a datetime.}
\usage{
to_datetime(.x)
}
\arguments{
\item{.x}{An object to convert to a datetime.}
}
\description{
Function to format an object into a datetime.
}
