% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{unbox_details}
\alias{unbox_details}
\title{unbox only elements that should be scalars}
\usage{
unbox_details(.x)
}
\arguments{
\item{.x}{list from collect_r_details()}
}
\description{
unbox only elements that should be scalars
}
