% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{whoami}
\alias{whoami}
\title{Function to get the user, or a mock user when the user isn't available.}
\usage{
whoami(session, .mock_user = "UNKNOWN")
}
\arguments{
\item{session}{The shiny session object.}

\item{.mock_user}{The "default" username to use if one is not available in session$user.}
}
\description{
Function to get the user, or a mock user when the user isn't available.
}
